requirements.txt
scikit-learn
flask
pandas
numpy
matplotlib
Pillow